third_party/blink/web_tests/external/wpt/permissions-policy/permissions-policy-opaque-origin.https-expected.txt

This is a testharness.js-based test.
[FAIL] permissions-policy-opaque-origin
  assert_equals: expected "fullscreen disabled in opaque-origin2.html" but got "fullscreen enabled in opaque-origin2.html"
Harness: the test ran to completion.
